backend/requirements.txt
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.1
comm==0.2.2
cssselect==1.3.0
debugpy==1.8.14
decorator==5.2.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
fake-useragent==2.2.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.4.0
lxml_html_clean==0.4.2
matplotlib-inline==0.1.7
nanoid==2.0.0
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyee==11.1.1
Pygments==2.19.1
PyJWT==2.9.0
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
requests-html==0.10.0
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==1.26.20
w3lib==2.3.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
